pyproject.toml
[project]
name = "soc-copilot"
version = "0.1.0"
description = "An AI investigation copilot that takes an alert and walks you through the analyst workflow."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.96.0",
    "httpx>=0.28.1",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]